pyproject.toml
[project]
name = "territories"
version = "0.0.3"
authors = [{ name="Timo", email="timothee@datapolitics.fr" }]
description = "Manage territories."
readme = "README.md"
requires-python = ">=3.10"


dependencies = [
  "networkx",
  "python-dotenv",
  "perfect_hash",
  "more-itertools",
  "psycopg2-binary"
]


[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"